# Optimize pool CPU