#### Description
<!--
Please describe the pull request here.
Also, if this is an issue/bug fix, please add the issue link for reference here.
-->
Optimize the CPU implementation of pool.

Benchmark code:

```
     tf.setBackend("cpu");
     const ns = [64, 128, 256, 512];
     const RUNS = 100;
     const filterSize = [3, 3];
     const strides = [1, 1];

    ns.forEach(n => {
      const x = tf.randomUniform([8, n, n, 3])
      
	// Warmup.
      tf.maxPool(x, filterSize, strides, 0);

      let res = null;
      const start = performance.now();
      for (let i = 0; i < RUNS; i++) {
        res = tf.maxPool(x, filterSize, strides, 0);
      }
      res.dataSync();
      const elapsed = (performance.now() - start) / RUNS;
      console.log(`N: ${n}: ${elapsed.toFixed(2)} ms`);
    });
```

Results before optimization:
```
N: 64: 83.81 ms
N: 128: 403.73 ms
N: 256: 1423.62 ms
N: 512: 6882.49 ms
```

Results after optimization:
```
 N: 64: 7.02 ms
 N: 128: 23.47 ms
 N: 256: 91.11 ms
 N: 512: 414.77 ms
```

---
<!-- Please do not delete this section -->
##### For repository owners only:

Please remember to apply all applicable tags to your pull request.
Tags: FEATURE, BREAKING, BUG, PERF, DEV, DOC, SECURITY

For more info see: https://github.com/tensorflow/tfjs/blob/master/DEVELOPMENT.md

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-core/1374)
<!-- Reviewable:end -->